--- front/src/App.js
@@ -1,9 +1,0 @@
-import Input from './input/input';
-import './App.css';
-
-
-export default () => (
-    <div className="center">
-        <Input />
-    </div>
-);
