+++ front/src/views/App.css
@@ -1,0 +1,6 @@
+.center {
+    text-align: center;
+    margin: auto;
+    width: 50%;
+    padding: 10px;
+}
