+++ front/src/views/input/input.css
@@ -3,3 +3,8 @@
     border-left-color: transparent;
     border-right-color: transparent;
 }
+
+.Hr{
+    width:70%;
+    margin-bottom: 30px;
+}
