+++ front/src/views/weatherView/weatherView.js
@@ -1,0 +1,1 @@
+export default () => <div></div>;
